Delete shit
Deleted a backup because i have a backup somewhere else so dont need also you know it being on github doesnt help. Deleted the asl sense im not using it. deleted the cheat table sense i decided i dont want to use the scripts in it to get pointers for the thing. Need to redo my personal cheat table so all the pointers are allign with the project and on the asl so they are allign